requirements.txt
asgiref
certifi
charset-normalizer
crispy-bootstrap5
dj-database-url
dj-rest-auth
Django
django-allauth
django-cors-headers
django-crispy-forms
django-environ
django-filter
djangorestframework
idna
Markdown
MouseInfo
pillow
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
PyScreeze
python-dateutil
pytweening
requests
six
sqlparse
typing_extensions
tzdata
urllib3
